+++ crates/nana-ui-scene/src/scene/primitives/overlays.rs
@@ -363,7 +363,7 @@ pub(super) fn build(context: &GeometryPaintContext<'_>, emit: &mut impl FnMut(Sc
                             z_index: node.z_index,
                             document_order: node_order,
                         },
-                        10u8.saturating_add(index as u8),
+                        10u64.saturating_add(index as u64),
                         scene_rect(option.bounds),
                         VisualQuadStyle {
                             background: Some(background),
@@ -377,7 +377,7 @@ pub(super) fn build(context: &GeometryPaintContext<'_>, emit: &mut impl FnMut(Sc
                     emit(ScenePrimitive {
                         id: PrimitiveId {
                             node: id,
-                            slot: 80u8.saturating_add(index as u8),
+                            slot: 80u64.saturating_add(index as u64),
                         },
                         node: id,
                         bounds: scene_rect(icon_bounds),
@@ -394,7 +394,7 @@ pub(super) fn build(context: &GeometryPaintContext<'_>, emit: &mut impl FnMut(Sc
                 }
                 emit(component_text_primitive(
                     id,
-                    40u8.saturating_add(index as u8),
+                    40u64.saturating_add(index as u64),
                     &option.label,
                     TextHorizontalAlignment::Start,
                     true,
